Split out product slug exports
